chore: better charts files
