+++ web/apps/benchmarks-and-reports/src/app/(benchmarks-and-reports)/benchmarks/page.tsx
@@ -1,8 +1,6 @@
 import type { Metadata } from "next";
-import { Suspense } from "react";
-import { SuspenseLoader } from "shared/client/components/suspense-loader";
 import { BENCHMARKS_DESCRIPTION } from "../[version]/_utils/constants";
-import { Charts } from "./_components/charts";
+import { Benchmarks } from "./_components/benchmarks";
 
 export const metadata: Metadata = {
   title: "Benchmarks",
@@ -21,9 +19,7 @@ export const metadata: Metadata = {
 export default function BenchmarksPage() {
   return (
     <div className="container max-w-screen-3xl">
-      <Suspense fallback={<SuspenseLoader />}>
-        <Charts />
-      </Suspense>
+      <Benchmarks />
     </div>
   );
 }
